Pulling log/syslogger test helpers into a separate file.
